exercise_consumer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-cloud-storage==2.17.0
